# openapi.yaml
openapi: 3.1.0
info:
  title: PFF Fantasy League API
  version: "1.0.0"
  description: |
    REST interface for the PFF Fantasy League toolkit. All requests must include
    the header `X-API-Key` matching the `PFF_API_KEY` environment variable.
servers:
  - url: https://pff-fantasy-analyzer.onrender.com
    description: Production
  - url: http://localhost:8000
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    LeaderboardEntry:
      type: object
      properties:
        team: {type: string}
        value: {type: number}
    StarterDetail:
      type: object
      properties:
        name: {type: string}
        position: {type: string, nullable: true}
        csvName: {type: string, nullable: true}
        rank: {type: integer, nullable: true}
        posRank: {type: integer, nullable: true}
        projection: {type: number, nullable: true}
        vor: {type: number, nullable: true}
    BenchDetail:
      type: object
      properties:
        name: {type: string}
        position: {type: string, nullable: true}
        rank: {type: integer, nullable: true}
        posRank: {type: integer, nullable: true}
        projection: {type: number, nullable: true}
        vor: {type: number, nullable: true}
        benchScore: {type: number, nullable: true}
        oVAR: {type: number, nullable: true}
    TeamDetail:
      type: object
      properties:
        team: {type: string}
        starters:
          type: array
          items: {$ref: '#/components/schemas/StarterDetail'}
        bench:
          type: array
          items: {$ref: '#/components/schemas/BenchDetail'}
        benchLimit: {type: integer, nullable: true}
...
    WaiverCandidate:
      type: object
      properties:
        name: {type: string}
        position: {type: string, nullable: true}
        team: {type: string, nullable: true}
        rank: {type: integer, nullable: true}
        posRank: {type: integer, nullable: true}
        projPoints: {type: number, nullable: true}
        vor: {type: number, nullable: true}
        benchScore: {type: number, nullable: true}
        oVAR: {type: number, nullable: true}
        needFactor: {type: number, nullable: true}
    WaiverListResponse:
      type: object
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/WaiverCandidate'}
        total: {type: integer}
        limit: {type: integer}
        offset: {type: integer}
        positionFilter: {type: string, nullable: true}
        teamFilter: {type: string, nullable: true}
    WaiverChange:
      type: object
      properties:
        team: {type: string}
        adds:
          type: array
          items: {type: string}
        drops:
          type: array
          items: {type: string}
      required: [team]
    WaiverRecommendRequest:
      type: object
      properties:
        changes:
          type: array
          items: {$ref: '#/components/schemas/WaiverChange'}
        includeDetails: {type: boolean, default: true}
        benchLimit: {type: integer, nullable: true}
      required: [changes]
    WaiverTeamResult:
      type: object
      properties:
        team: {type: string}
        baseline: {type: number}
        postChange: {type: number}
        delta: {type: number}
    WaiverRecommendResponse:
      type: object
      properties:
        evaluatedAt: {type: string, format: date-time}
        teams:
          type: array
          items: {$ref: '#/components/schemas/WaiverTeamResult'}
        combinedScores:
          type: object
          additionalProperties: {type: number}
        leaderboards:
          type: object
          additionalProperties:
            type: array
            items: {$ref: '#/components/schemas/LeaderboardEntry'}
        details:
          type: array
          items: {$ref: '#/components/schemas/TeamDetail'}
          nullable: true
...
paths:
...
  /waivers/candidates:
    get:
      tags: [waivers]
      summary: List waiver candidates
      operationId: waiverCandidates
      x-openai-isConsequential: false
      parameters:
        - name: team
          in: query
          description: Team context to weight positional needs
          schema: {type: string}
        - name: position
          in: query
          description: Filter by position
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, default: 25, minimum: 1, maximum: 200}
        - name: offset
          in: query
          schema: {type: integer, default: 0, minimum: 0}
      responses:
        '200':
          description: Ranked waiver targets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaiverListResponse'
  /waivers/recommend:
    post:
      tags: [waivers]
      summary: Evaluate add/drop scenarios
      operationId: waiverRecommend
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaiverRecommendRequest'
            example:
              changes:
                - team: JohnBurrows_School
                  adds: [Drake London]
                  drops: [George Pickens]
              includeDetails: true
              benchLimit: 2
      responses:
        '200':
          description: League impact of proposed moves
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaiverRecommendResponse'
...
security:
  - ApiKeyAuth: []